requirements.txt
SpeechRecognition
pyaudio
google-generativeai